+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.
 - Fix RetroFlag scripts to turn off cases (#2067)
 - Fix tearing effect on odroidgo2 (#2100)
 - Add DRM info in support archive (#2102)
+- Add pulseaudio info in support archive (#2118)
 
 ## [8.0-Electron]
 - Remove non-working shaders